opt: don't allocate string unless not already interned
Optimise string concatnation by allocating result string
only if the resultant string does not already exist in the
interened string table. This relies on the fact that if we
start hashing string `b` with hash of string `a` as the starting
point, we end up with the hash of the string `a+b`.